+++ srcs/lin_alg/vec3_operations.c
@@ -6,13 +6,13 @@
 /*   By: dshatilo <dshatilo@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/05/27 09:56:19 by dshatilo          #+#    #+#             */
-/*   Updated: 2024/05/29 12:47:09 by dshatilo         ###   ########.fr       */
+/*   Updated: 2024/06/24 21:31:16 by dshatilo         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "lin_alg.h"
+#include "minirt.h"
 
-t_vec3	vec3(double x, double y, double z)
+t_vec3	vec3(float x, float y, float z)
 {
 	t_vec3	vec3or;
 
@@ -52,7 +52,7 @@ t_vec3	vec3_prod(t_vec3 a, t_vec3 b)
 	return (prod);
 }
 
-t_vec3	vec3_scale(double coef, t_vec3 v)
+t_vec3	vec3_scale(float coef, t_vec3 v)
 {
 	t_vec3	new;
