+++ srcs/lin_alg/vec3_operations2.c
@@ -6,13 +6,13 @@
 /*   By: dshatilo <dshatilo@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/05/27 12:20:51 by dshatilo          #+#    #+#             */
-/*   Updated: 2024/05/29 12:48:08 by dshatilo         ###   ########.fr       */
+/*   Updated: 2024/06/24 21:31:25 by dshatilo         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "lin_alg.h"
+#include "minirt.h"
 
-double	vec3_norm(t_vec3 vect)
+float	vec3_norm(t_vec3 vect)
 {
 	return (sqrt(dot2(vect)));
 }
@@ -20,7 +20,7 @@ double	vec3_norm(t_vec3 vect)
 t_vec3	vec3_unit(t_vec3 vect)
 {
 	t_vec3	unit;
-	double	norm;
+	float	norm;
 
 	norm = vec3_norm(vect);
 	unit.x = vect.x / norm;
@@ -29,19 +29,19 @@ t_vec3	vec3_unit(t_vec3 vect)
 	return (unit);
 }
 
-double	vec3_cos(t_vec3 a, t_vec3 b)
+float	vec3_cos(t_vec3 a, t_vec3 b)
 {
 	return (dot(a, b) / (vec3_norm(a) * vec3_norm(b)));
 }
 
-double	vec3_sin(t_vec3 a, t_vec3 b)
+float	vec3_sin(t_vec3 a, t_vec3 b)
 {
 	return (sqrt(1 - pow(vec3_cos(a, b), 2)));
 }
 
-double	distance(t_vec3 v1, t_vec3 v2)
+float	distance(t_vec3 v1, t_vec3 v2)
 {
-	double	d;
+	float	d;
 
 	d = sqrt(pow(v2.x - v1.x, 2) + pow(v2.y - v1.y, 2) + pow(v2.z - v1.z, 2));
 	return (d);
